UPSTREAM: tracing: Add array printing helper
If a trace event contains an array, there is currently no standard
way to format this for text output.  Drivers are currently hacking
around this by a) local hacks that use the trace_seq functionailty
directly, or b) just not printing that information.  For fixed size
arrays, formatting of the elements can be open-coded, but this gets
cumbersome for arrays of non-trivial size.

These approaches result in non-standard content of the event format
description delivered to userspace, so userland tools needs to be
taught to understand and parse each array printing method
individually.

This patch implements a __print_array() helper that tracepoint
implementations can use instead of reinventing it.  A simple C-style
syntax is used to delimit the array and its elements {like,this}.

So that the helpers can be used with large static arrays as well as
dynamic arrays, they take a pointer and element count: they can be
used with __get_dynamic_array() for use with dynamic arrays.
Link: http://lkml.kernel.org/r/1422449335-8289-2-git-send-email-javi.merino@arm.com

Cc: Ingo Molnar <mingo@redhat.com>
Signed-off-by: Dave Martin <Dave.Martin@arm.com>
Signed-off-by: Javi Merino <javi.merino@arm.com>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>
(cherry picked from commit 6ea22486ba46bcb665de36514094d74575cd1330)
Signed-off-by: Kevin Hilman <khilman@linaro.org>

BUG=chrome-os-partner:42804
TEST=`emerge-oak chromeos-kernel-3_18`

Change-Id: Ib073074b670f6977420d247e9a1a30e26385d80a
Signed-off-by: Dawei.Chien <dawei.chien@mediatek.com>
Reviewed-on: https://chromium-review.googlesource.com/283821
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Wei-Ning Huang <wnhuang@chromium.org>
Commit-Queue: Wei-Ning Huang <wnhuang@chromium.org>