CHROMIUM: wireless: allow self-managed regulatory
This is a very limited backport of some cfg80211 regulatory
infrastructure - which will rely on having more of the code
in the driver/mac80211 - of the wiphy specific regulatory
handling from upstream.

Backporting all of the upstream code would have been a huge
change that affects all drivers, this change is specifically
tailored to have no impact on drivers that don't set the new
WIPHY_FLAG_SELF_MANAGED_REG flag.

BUG=N/A
TEST=run with new iwl7000 using this functionality

Change-Id: I4c5cbaffba60870e287cc18548ee04b677ae8835
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/273856
Reviewed-by: Paul Stewart <pstew@chromium.org>