CHROMIUM: vb2: Don't reuse USERPTR buffers.
The logic in vb2 USERPTR implementation assumes that userspace will always
use the same V4L2 buffer index with one virtual address for the duration
of its usage. Based on that, it does not release a pinned buffer, even when
it is returned to userspace, until a different address is queued for the
same index, because it assumes that when that happens, the userspace has
changed the working set of buffers and the old set is not in use anymore.

However, if userspace does not maintain a constant address to index
relationship, vb2 will release the previous buffer pinned for that index
at the time of queuing the new one, which is after the old buffer has been
returned to the userspace. That may be destructive if that buffer is
already in use elsewhere at that time, as userspace expects that kernel
disclaimed the old buffer before it was dequeued.

Signed-off-by: Pawel Osciak <posciak@chromium.org>

TEST=corruption test with V4L2VEA
BUG=chrome-os-partner:35860

Original-Change-Id: I10c67f89aaea7fb5247a223aad4ac9728817af7a
Reviewed-on: https://chromium-review.googlesource.com/246970
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Tested-by: Pawel Osciak <posciak@chromium.org>
Commit-Queue: Pawel Osciak <posciak@chromium.org>
(cherry picked from commit f0ce3a1a20ebc892ab478b127b680ba9a305adbb)
Signed-off-by: Pawel Osciak <posciak@chromium.org>

Change-Id: I836fe0165ec08a78d6b9cb91e03e994163035e23
Reviewed-on: https://chromium-review.googlesource.com/302513
Commit-Ready: Pawel Osciak <posciak@chromium.org>
Tested-by: Pawel Osciak <posciak@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>