FROMLIST: init: Export name_to_dev_t and mark it const
name_to_dev_t is useful in other pieces of code to initialize rootfs.
In the case of dm, the code is sometimes built in a module and other
times used to construct the rootfs; therefore it must be exported
as a symbol.

BUG=chromium:442794
TEST=Boots with verity from Urara

Signed-off-by: Dan Ehrenberg <dehrenberg@chromium.org>
(am from https://www.redhat.com/archives/dm-devel/2015-February/msg00041.html)
Change-Id: Icc07a860aec754b5b960fce793fa55d7efc3d8ab
Reviewed-on: https://chromium-review.googlesource.com/255760
Reviewed-by: Kees Cook <keescook@chromium.org>