UPSTREAM: ALSA: hda - Use regmap for command verb caches, too
Like the previous patches, this patch converts also to the regmap, at
this time, the cached verb writes are the target.  But this conversion
needs a bit more caution than before.

- In the old code, we just record any verbs as is, and restore them at
  resume.  For the regmap scheme, this doesn't work, since a few verbs
  like AMP or DIGI_CONVERT are asymmetrical.  Such verbs are converted
  either to the dedicated function (snd_hda_regmap_xxx_amp()) or
  changed to the unified verb.

- Some verbs have to be declared as vendor-specific ones before
  accessing via regmap.

Also, the minor optimization with codec->cached_write flag is dropped
in a few places, as this would confuse the operation.  Further
optimizations will be brought in the later patches, if any.

This conversion ends up with a drop of significant amount of codes,
mostly the helper codes that are no longer used.

Signed-off-by: Takashi Iwai <tiwai@suse.de>
Link: git://git.kernel.org/pub/scm/linux/kernel/git/next/linux-next.git
br: master
(cherry picked from commit a551d91473e5e3a591f6fe86ac5a5fb460c3f96a)
Signed-off-by: Vedang Patel <vedang.patel@intel.com>

BUG=None
TEST=with the whole series applied, the audio subsystem is stable.
TEST=there are no regressions observed on BSW.

Change-Id: Iea5ecaeed7a39f84abd09ea2a3406279b7f5ed81