FROMLIST: mtd: part: Create the master device node when partitioned
For many use cases, it helps to have a device node for the entire
MTD device as well as device nodes for the individual partitions.
For example, this allows querying the entire device's properties.
A common idiom is to create an additional partition which spans
over the whole device.

This patch makes a config option, CONFIG_MTD_PARTITIONED_MASTER,
which makes the master partition present even when the device is
partitioned. This isn't turned on by default since it presents
a backwards-incompatible device numbering.

The patch also makes the parent of a partition device be the master,
if the config flag is set, now that the master is a full device.

BUG=chromium:442794
TEST=booted on urara and added and removed partitions on mtd0;
observed that mtd0 had a name derived from the whole device.
CQ-DEPEND=CL:257690

Signed-off-by: Dan Ehrenberg <dehrenberg@chromium.org>
(am from https://patchwork.ozlabs.org/patch/449616/)
Change-Id: I0849c98185e7231457cc48f589baf96d8d989da4
Reviewed-on: https://chromium-review.googlesource.com/261070