FROMLIST: ARM: mm: Do not invoke OOM for higher order IOMMU DMA allocations
IOMMU should be able to use single pages as well as bigger blocks, so if
higher order allocations fail, we should not affect state of the system,
with events such as OOM killer, but rather fall back to order 0
allocations.

This patch changes the behavior of ARM IOMMU DMA allocator to use
__GFP_NORETRY, which bypasses OOM invocation, for orders higher than
zero and, only if that fails, fall back to normal order 0 allocation
which might invoke OOM killer.

Signed-off-by: Tomasz Figa <tfiga@chromium.org>

BUG=chrome-os-partner:37755
TEST=repeated video_VideoSeek with 1.6G on tmpfs and two cat /dev/zero > /dev/null running for additional memory fragmentation on veyron_jerry

Change-Id: Id2a91db0ff89add7c3ef1bb8156d75c652fb6e44
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
(am from https://patchwork.kernel.org/patch/6070801/)
Reviewed-on: https://chromium-review.googlesource.com/260194
Reviewed-by: Douglas Anderson <dianders@chromium.org>