CHROMIUM: memory: tegra: Add API needed by the EMC driver
The EMC driver needs to know the number of external memory devices and
also needs to update the EMEM configuration based on the new rate of the
memory bus.

To know how to update the EMEM config, looks up the values of the burst
regs in the DT, for a given timing.

Change-Id: Ifa648625154dd14f1b18eeb026f5344adb9365fa
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Signed-off-by: Tomeu Vizoso <tomeu.vizoso@collabora.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>
Reviewed-on: https://chrome-internal-review.googlesource.com/213936
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Tested-by: Andrew Chew <achew@nvidia.com>
Commit-Queue: Andrew Chew <achew@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/267613
Reviewed-by: David Riley <davidriley@chromium.org>
Tested-by: David Riley <davidriley@chromium.org>
Commit-Queue: David Riley <davidriley@chromium.org>